backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
google-generativeai==0.5.4
sentence-transformers==2.7.0
chromadb==0.4.24
numpy==1.26.4
httpx==0.26.0
python-multipart==0.0.9
